modularize token response creation
